[parallel] include ranged.hh from EXADUNE if available, use our copy else
